requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-html>=4.0.0
pytest-xdist>=3.5.0
PyJWT>=2.8.0
pyyaml>=6.0.1
openai>=1.12.0
python-dotenv>=1.0.0
